Revert "Fix/document type page"
