--- core/document-types/page.php
@@ -1,31 +1,0 @@
-<?php
-namespace Elementor\Core\DocumentTypes;
-
-if ( ! defined( 'ABSPATH' ) ) {
-	exit; // Exit if accessed directly
-}
-
-class Page extends Base {
-	public static function get_properties() {
-		$properties = parent::get_properties();
-
-		$properties['cpt'] = [ 'page' ];
-
-		return $properties;
-	}
-
-	/**
-	 * @access public
-	 */
-	public function get_name() {
-		return 'wp-page';
-	}
-
-	/**
-	 * @access public
-	 * @static
-	 */
-	public static function get_title() {
-		return __( 'Page', 'elementor' );
-	}
-}
